Refactor mail service to use Kafka for email events
- Updated MailService to send verification and invite emails via Kafka.
- Removed gRPC mail client and proto definitions.

+++ mail-service/internal/services/mail.service.go
@@ -45,7 +45,7 @@ func (s *MailService) SendVerifyEmail(to, username, code string, expiresMinutes 
 	return s.sendMail(to, "Logtheus email verification", body)
 }
 
-func (s *MailService) SendInviteEmail(to, inviteeName, referrer, projectName, code string, expiresMinutes uint8) error {
+func (s *MailService) SendInviteEmail(to, inviteeName, referrer, projectName, code string, expiresMinutes uint16) error {
 	url := fmt.Sprintf("%s/accept-invite/%s", s.domain, code)
 	data := &types.InviteEmailData{InviteeName: inviteeName, Referrer: referrer, ProjectName: projectName, InviteLink: url, ExpiresIn: expiresMinutes}
