+++ mail-service/internal/types/template.types.go
@@ -12,5 +12,5 @@ type InviteEmailData struct {
 	Referrer    string
 	ProjectName string
 	InviteLink  string
-	ExpiresIn   uint8
+	ExpiresIn   uint16
 }
